+++ tests/test_commands.py
@@ -34,6 +34,7 @@ class CommandTest(unittest.TestCase):
             'FairSwap',
             'FairSwap-Reusable',
             'Delgado',
+            'Delgado-Library',
             'Delgado-Reusable',
             'SimplePayment',
             'SimplePayment-prepaid',
@@ -107,6 +108,12 @@ class CommandTest(unittest.TestCase):
 
     def test_run_delgado_reusable(self):
         p = subprocess.Popen(['bdtsim', 'run', 'Delgado-Reusable', 'PyEVM'],
+                             stdout=subprocess.PIPE)
+        out, err = p.communicate()
+        self.assertEqual(p.returncode, 0)
+
+    def test_run_delgado_library(self):
+        p = subprocess.Popen(['bdtsim', 'run', 'Delgado-Library', 'PyEVM'],
                              stdout=subprocess.PIPE)
         out, err = p.communicate()
         self.assertEqual(p.returncode, 0)